Run firefox as non-root
Linux only.

Due to https://bugzilla.mozilla.org/show_bug.cgi?id=1323302 the firefox
executable can no longer be run by root.  So when chef tries to run
`firefox -v` it gets an error instead of the version.

This change adds a user to run the command as and then deletes the user.